connected dashboard to the backend server
